# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Barber Booking System API for Kosovo
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [products]
      summary: List orders
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: shopId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"

    post:
      operationId: createOrder
      tags: [products]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [products]
      summary: Get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

    patch:
      operationId: updateOrder
      tags: [products]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # PAYMENTS
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /dashboard/owner-stats:
    get:
      operationId: getOwnerStats
      tags: [dashboard]
      summary: Get barbershop owner dashboard stats
      parameters:
        - name: shopId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerStats"
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent platform activity
      parameters:
        - name: shopId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/appointments-by-day:
    get:
      operationId: getAppointmentsByDay
      tags: [dashboard]
      summary: Appointment counts grouped by day
      parameters:
        - name: shopId
          in: query
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Appointments by day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayCount"
  /dashboard/revenue-by-month:
    get:
      operationId: getRevenueByMonth
      tags: [dashboard]
      summary: Revenue grouped by month
      parameters:
        - name: shopId
          in: query
          schema:
            type: integer
        - name: months
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue by month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthRevenue"
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, owner, user]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, shopId, name, price, createdAt]
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        stock:
          type: integer
        category:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Order:
      type: object
      required: [id, userId, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, paid, shipped, delivered, cancelled]
        totalAmount:
          type: number
        stripeSessionId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required: [id, orderId, productId, quantity, unitPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        product:
          $ref: "#/components/schemas/Product"
    OrderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [paid, shipped, delivered, cancelled]
    OrderList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer

    # PAYMENT
...
    Notification:
      type: object
      required: [id, userId, title, message, type, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [booking_confirmed, booking_cancelled, booking_reminder, order_update, general]
        isRead:
          type: boolean
        relatedId:
          type: ["integer", "null"]
        relatedType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    NotificationList:
      type: object
      required: [data, total, unreadCount]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer

    # DASHBOARD
...
    OwnerStats:
      type: object
      required: [totalAppointments, confirmedAppointments, totalRevenue, totalBarbers, totalProducts]
      properties:
        totalAppointments:
          type: integer
        confirmedAppointments:
          type: integer
        cancelledAppointments:
          type: integer
        totalRevenue:
          type: number
        todayAppointments:
          type: integer
        totalBarbers:
          type: integer
        totalProducts:
          type: integer
        rating:
          type: ["number", "null"]
        subscriptionActive:
          type: boolean
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [appointment_booked, appointment_confirmed, appointment_cancelled, shop_registered, product_ordered]
        description:
          type: string
        userName:
          type: ["string", "null"]
        shopName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    DayCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer
    MonthRevenue:
      type: object
      required: [month, revenue]
      properties:
        month:
          type: string
        revenue:
          type: number